Display error message on fail plugin vim install

+++ editors/vim/plugins/racer/install.sh
@@ -1,17 +1,18 @@
 #!/usr/bin/env bash
 install_plugin "${HOME}/.vim/bundle/vim-racer" https://github.com/racer-rust/vim-racer.git master
 
-FILE_PLUGIN="${HOME}/.vim/plugin/rust_racer.vim"
+if [ $? -eq 0 ]; then
+  FILE_PLUGIN="${HOME}/.vim/plugin/rust_racer.vim"
 
-if [ -f "${FILE_PLUGIN}" ]; then
-  rm -rf "${FILE_PLUGIN}"
-fi
+  if [ -f "${FILE_PLUGIN}" ]; then
+    rm -rf "${FILE_PLUGIN}"
+  fi
 
-ln -s "${HOME}/.vim/bundle/vim-racer/ftplugin/rust_racer.vim" "${FILE_PLUGIN}"
+  ln -s "${HOME}/.vim/bundle/vim-racer/ftplugin/rust_racer.vim" "${FILE_PLUGIN}"
 
-CHANNEL=$(cat "${TMP_RUST_CHANNEL}")
+  CHANNEL=$(cat "${TMP_RUST_CHANNEL}")
 
-RUST_SYS_ROOT=$(rustup run "${CHANNEL}" rustc --print sysroot)
+  RUST_SYS_ROOT=$(rustup run "${CHANNEL}" rustc --print sysroot)
 
 # Enable plugin
 cat <<EOF >>${HOME}/.vimrc
@@ -26,3 +27,6 @@ let \$RUST_SRC_PATH="${RUST_SYS_ROOT}/lib/rustlib/src/rust/src"
 "-------------------------------------------------------------------------------
 
 EOF
+else
+  exit 1
+fi
